# refactor: improve readability of benchmarking output & some DRY code optimizations.