Thanks @imetandy for the work done on this hashing tool.

I wrote this pull request to introduce some simple improvements:
- Adds a hashes per second metric in output.
- Adds a `measure_hash_speed` that accepts a closure associated with the corresponding hashing algorithm.

New out put looks like this:

```sh
SHA256:
SHA256 - 8 bytes: 256551.80 hash/sec, 0.000003898 seconds per hash
SHA256 - 16 bytes: 260398.00 hash/sec, 0.000003840 seconds per hash
SHA256 - 32 bytes: 267890.75 hash/sec, 0.000003733 seconds per hash
SHA256 - 64 bytes: 169188.71 hash/sec, 0.000005911 seconds per hash
SHA256 - 128 bytes: 126362.77 hash/sec, 0.000007914 seconds per hash

SHA512:
SHA512 - 8 bytes: 198122.61 hash/sec, 0.000005047 seconds per hash
SHA512 - 16 bytes: 199077.92 hash/sec, 0.000005023 seconds per hash
SHA512 - 32 bytes: 202632.53 hash/sec, 0.000004935 seconds per hash
SHA512 - 64 bytes: 206416.24 hash/sec, 0.000004845 seconds per hash
SHA512 - 128 bytes: 141821.30 hash/sec, 0.000007051 seconds per hash

BLAKE3:
BLAKE3 - 8 bytes: 556142.02 hash/sec, 0.000001798 seconds per hash
BLAKE3 - 16 bytes: 566290.63 hash/sec, 0.000001766 seconds per hash
BLAKE3 - 32 bytes: 561938.87 hash/sec, 0.000001780 seconds per hash
BLAKE3 - 64 bytes: 553985.75 hash/sec, 0.000001805 seconds per hash
BLAKE3 - 128 bytes: 309082.72 hash/sec, 0.000003235 seconds per hash
```